Fix misc bugs
